Refactor to leverage persistent layout
