+++ next.config.js
@@ -1,6 +1,6 @@
 const nextra = require("nextra");
 
-module.exports = nextra("./components/layout.js")({
+module.exports = nextra("./components/layout-nextra.js")({
   reactStrictMode: true,
   unstable_staticImage: true,
 });
