+++ pages/index.mdx
@@ -1,4 +1,4 @@
-import { DESCRIPTION } from "../components/layout";
+import { DESCRIPTION } from '../components/layout';
 
 # Hi. 👋🏻 <br /> I’m David Calhoun.
